+++ core-ng/src/main/java/core/framework/internal/log/appender/KafkaAppender.java
@@ -84,11 +84,11 @@ public final class KafkaAppender implements LogAppender {
                     ProducerConfig.ACKS_CONFIG, "0",                                        // no acknowledge to maximize performance
                     ProducerConfig.CLIENT_ID_CONFIG, "log-forwarder",                       // if not specify, kafka uses producer-${seq} name, also impact jmx naming
                     ProducerConfig.COMPRESSION_TYPE_CONFIG, CompressionType.SNAPPY.name,
-                    ProducerConfig.DELIVERY_TIMEOUT_MS_CONFIG, 60 * 1000,                   // 60s, type is INT
+                    ProducerConfig.DELIVERY_TIMEOUT_MS_CONFIG, 60_000,                      // 60s, type is INT
                     ProducerConfig.LINGER_MS_CONFIG, 50L,
                     ProducerConfig.RECONNECT_BACKOFF_MS_CONFIG, 500L,                       // longer backoff to reduce cpu usage when kafka is not available
-                    ProducerConfig.RECONNECT_BACKOFF_MAX_MS_CONFIG, 5L * 1000,              // 5s
-                    ProducerConfig.MAX_BLOCK_MS_CONFIG, 30L * 1000);                        // 30s, metadata update timeout, shorter than default, to get exception sooner if kafka is not available
+                    ProducerConfig.RECONNECT_BACKOFF_MAX_MS_CONFIG, 5_000L,                 // 5s
+                    ProducerConfig.MAX_BLOCK_MS_CONFIG, 30_000L);                           // 30s, metadata update timeout, shorter than default, to get exception sooner if kafka is not available
             var serializer = new ByteArraySerializer();
             var producer = new KafkaProducer<>(config, serializer, serializer);
             producerMetrics.set(producer.metrics());
@@ -100,7 +100,9 @@ public final class KafkaAppender implements LogAppender {
 
     @Override
     public void append(ActionLogMessage message) {
-        records.add(new ProducerRecord<>(LogTopics.TOPIC_ACTION_LOG, actionLogWriter.toJSON(message)));     // not specify message key for sticky partition
+        // not specify message key for sticky partition, StickyPartitionCache will be used if key is null
+        // refer to org.apache.kafka.clients.producer.internals.DefaultPartitioner.partition
+        records.add(new ProducerRecord<>(LogTopics.TOPIC_ACTION_LOG, actionLogWriter.toJSON(message)));
     }
 
     @Override